+++ app/src/main/java/com/example/weather/detail/HourlyWeatherInformation.kt
@@ -3,7 +3,7 @@ package com.example.weather.detail
 enum class TimeStage { Now, Normal, Sunset, Sunrise }
 
 data class HourlyWeatherInformation(
-    val temperature: Int?,
+    val temperature: Double?,
     val icon: String,
     val chanceOfRain: Double?,
     val chanceOfSnow: Double?,
